first plane animation refactored animation code added resource loading manager

+++ project/source/engine/vdptile_manager.h
@@ -10,6 +10,6 @@ void VDPTileManager_Init(void);
 // returns the sprite arearelative vdp index (ie returns 0 for the first sprite tile, not 256)
 u8 VDPTileManager_LoadSpriteTiles(const u8* tileData, u8 tileCount); 
 
-u16 VDPTileManager_LoadBackgroundTileset(const u8* tileData, u8 tileCount);
+u16 VDPTileManager_LoadBackgroundTileset(const u8* tileData, u16 tileCount);
 
 #endif
